# Add optimized C implementation for `secp256k1` curve.