--- a/crypto/ec/ec_curve.c
+++ b/crypto/ec/ec_curve.c
@@ -2935,7 +2935,12 @@ static const ec_list_element curve_list[] = {
     {NID_secp224r1, &_EC_NIST_PRIME_224.h, 0,
      "NIST/SECG curve over a 224 bit prime field"},
 # endif
-    {NID_secp256k1, &_EC_SECG_PRIME_256K1.h, 0,
+    {NID_secp256k1, &_EC_SECG_PRIME_256K1.h,
+# if !defined(OPENSSL_NO_EC_NISTP_64_GCC_128)
+    EC_GFp_secp256k1_method,
+# else
+    0,
+#endif
      "SECG curve over a 256 bit prime field"},
     /* SECG secp256r1 is the same as X9.62 prime256v1 and hence omitted */
     {NID_secp384r1, &_EC_NIST_PRIME_384.h,