… crypto/ec/ecp_secp256k1.c
… crypto/ec/ecp_secp256k1_table.h