This Pull Request adds optimized C code on platforms supporting `uint128_t` type for the elliptic curve `secp256k1`.
It also adds `secp256k1` to the list of benchmarked curves.

The performance improvement is ~10x for ECDSA signature and  and ~4x for verification:
```
Before:
$ ./apps/openssl speed ecdsa
[...]
256 bits ecdsa (secp256k1)   0.0002s   0.0002s   4784.1   5074.3
[...]

After:
$ ./apps/openssl speed ecdsa
[...]
 256 bits ecdsa (secp256k1)   0.0000s   0.0000s  47047.1  21591.0
[...]
```

Fixes #23524

<!--
Thank you for your pull request. Please review these requirements:

Contributors guide: https://github.com/openssl/openssl/blob/master/CONTRIBUTING.md

Other than that, provide a description above this comment if there isn't one already

If this fixes a GitHub issue, make sure to have a line saying 'Fixes #XXXX' (without quotes) in the commit message.
-->

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->
- [x] documentation is added or updated
- [x] tests are added or updated