@MichaelA-Fireblocks 2025-01-16
@sipa Indeed. I'll update the code in this PR to match the most recent modifications that were made into [https://github.com/bitcoin-core/secp256k1](https://github.com/bitcoin-core/secp256k1)