@kroeckx 2025-01-17
There are cases where variable time is not a problem, but we don't want to add such implementations.