@MichaelA-Fireblocks 2025-09-25
Hello @sipa 
> If there is interest in this, I'm happy to take a look at the CLA, and ping other significant contributors.

Did you get a chance to look at the CLA thing ?
Thanks